+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/CustomerSheet/CustomerSheet+PaymentMethodAvailability.swift
@@ -54,3 +54,36 @@ extension Array where Element == STPPaymentMethodType {
         return pmTypes
     }
 }
+
+extension CustomerSheet {
+    /// Given a list of paymentMethodTypes, return an array of corresponding deduped list of STPPaymentMethodType that are supported
+    /// within customer sheet.  If any unsupported payment method types are passed in, return an error.
+    static func customerSheetSupportedPaymentMethodTypes(_ paymentMethodTypes: [String]) -> Result<[STPPaymentMethodType]?, Error> {
+        guard !paymentMethodTypes.isEmpty else {
+            return .success(nil)
+        }
+        var unsupportedPMs: [String] = []
+        var unsupportedPMsSet: Set<String> = []
+        var validPMs: [STPPaymentMethodType] = []
+        var validPMsSet: Set<STPPaymentMethodType> = []
+
+        for paymentMethodType in paymentMethodTypes {
+            let stpPaymentMethodType = STPPaymentMethod.type(from: paymentMethodType)
+            if !CustomerSheet.supportedPaymentMethods.contains(where: { $0 == stpPaymentMethodType }) {
+                if !unsupportedPMsSet.contains(paymentMethodType) {
+                    unsupportedPMsSet.insert(paymentMethodType)
+                    unsupportedPMs.append(paymentMethodType)
+                }
+            } else {
+                if !validPMsSet.contains(stpPaymentMethodType) {
+                    validPMsSet.insert(stpPaymentMethodType)
+                    validPMs.append(stpPaymentMethodType)
+                }
+            }
+        }
+        guard unsupportedPMs.isEmpty else {
+            return .failure(CustomerSheetError.unsupportedPaymentMethodType(paymentMethodTypes: unsupportedPMs))
+        }
+        return .success(validPMs)
+    }
+}