+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/CustomerSheet/CustomerSheetPaymentMethodAvailabilityTests.swift
@@ -48,6 +48,77 @@ class CustomerSheetPaymentMethodAvailabilityTests: XCTestCase {
 
         XCTAssertEqual(sut, [.card])
     }
+    func testCustomerSheetSupportedPaymentMethodTypes_card() {
+        let input = ["card"]
+        guard case .success(let result) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              let result else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(result, [.card])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_AllValid() {
+        let input = ["card", "us_bank_account", "sepa_debit"]
+        guard case .success(let result) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              let result else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(result, [.card, .USBankAccount, .SEPADebit])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_partialInvalid() {
+        let input = ["card", "us_bank_account", "llama_pay"]
+        guard case .failure(let err) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              case CustomerSheetError.unsupportedPaymentMethodType(let unsupported) = err else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(unsupported, ["llama_pay"])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_allInvalid() {
+        let input = ["llama_pay1", "llama_pay2"]
+        guard case .failure(let err) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              case CustomerSheetError.unsupportedPaymentMethodType(let unsupported) = err else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(unsupported, ["llama_pay1", "llama_pay2"])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_duplicate1() {
+        let input = ["card", "card"]
+        guard case .success(let result) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              let result else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(result, [.card])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_removeDupeInOrder() {
+        let input = ["card", "us_bank_account", "card"]
+        guard case .success(let result) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              let result else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(result, [.card, .USBankAccount])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_duplicate_withInvalid() {
+        let input = ["card", "card", "llama_pay"]
+        guard case .failure(let err) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input),
+              case CustomerSheetError.unsupportedPaymentMethodType(let unsupported) = err else {
+            XCTFail()
+            return
+        }
+        XCTAssertEqual(unsupported, ["llama_pay"])
+    }
+    func testCustomerSheetSupportedPaymentMethodTypes_empty() {
+        let input: [String] = []
+        guard case .success(let result) = CustomerSheet.customerSheetSupportedPaymentMethodTypes(input) else {
+            XCTFail()
+            return
+        }
+        XCTAssertNil(result)
+    }
 
     var mockCustomerAdapterWithSetupIntent: CustomerAdapter {
         return MockCustomerAdapter(mockedValue: true)
@@ -84,5 +155,8 @@ class MockCustomerAdapter: CustomerAdapter {
 
     func updatePaymentMethod(paymentMethodId: String, paymentMethodUpdateParams: StripePayments.STPPaymentMethodUpdateParams) async throws -> StripePayments.STPPaymentMethod {
         throw CustomerSheetError.unknown(debugDescription: "Not implemented")
+    }
+    var paymentMethodTypes: [String]? {
+        return nil
     }
 }